Feature/issue 1977 version kv join version kv (#2056)
This closed #1977

+++ src/Storages/ExternalStream/Kafka/KafkaSource.h
@@ -79,8 +79,8 @@ private:
     Int64 flush_interval_ms = 1000;
     Int64 last_flush_ms = 0;
 
-    UInt64 record_consume_batch_count = 1000;
-    Int64 record_consume_timeout = 100;
+    UInt32 record_consume_batch_count = 1000;
+    Int32 record_consume_timeout = 100;
 };
 
 }
